requirements.txt
numpy==1.17.0
pandas==1.0.1
scikit-learn==0.22.1
opencv-python==4.2.0.34
scipy==1.4.1
tensorflow-gpu==2.1.0